feat: allocate check max size
