+++ pkg/service/storage/server.go
@@ -3,6 +3,7 @@ package storage
 import (
 	"context"
 	"errors"
+	"fmt"
 	"time"
 
 	logging "github.com/ipfs/go-log/v2"
@@ -12,12 +13,15 @@ import (
 	"github.com/storacha/go-ucanto/principal"
 	"github.com/storacha/go-ucanto/server"
 	"github.com/storacha/go-ucanto/ucan"
+	"github.com/storacha/storage/pkg/capability"
 	"github.com/storacha/storage/pkg/capability/blob"
 	"github.com/storacha/storage/pkg/store"
 	"github.com/storacha/storage/pkg/store/allocationstore/allocation"
 )
 
 var log = logging.Logger("storage")
+
+const maxUploadSize = 127 * (1 << 25)
 
 func NewServer(id principal.Signer, storageService Service) (server.ServerView, error) {
 	return server.NewServer(
@@ -29,6 +33,8 @@ func NewServer(id principal.Signer, storageService Service) (server.ServerView, 
 				func(cap ucan.Capability[blob.AllocateCaveats], inv invocation.Invocation, iCtx server.InvocationContext) (blob.AllocateOk, receipt.Effects, error) {
 					log.Infof("%s z%s => %s", blob.AllocateAbility, cap.Nb().Blob.Digest.B58String(), cap.Nb().Space)
 					ctx := context.TODO()
+
+					// TODO: restrict invocation to authority (storacha service)
 
 					// check if we already have an allcoation for the blob in this space
 					allocs, err := storageService.Allocations().List(ctx, cap.Nb().Blob.Digest)
@@ -47,6 +53,13 @@ func NewServer(id principal.Signer, storageService Service) (server.ServerView, 
 							if !errors.Is(err, store.ErrNotFound) {
 								return blob.AllocateOk{}, nil, failure.FromError(err)
 							}
+						}
+					}
+
+					if cap.Nb().Blob.Size > maxUploadSize {
+						return blob.AllocateOk{}, nil, capability.Failure{
+							Name:    "BlobSizeOutsideOfSupportedRange",
+							Message: fmt.Sprintf("Blob of %d bytes, exceeds size limit of %d bytes", cap.Nb().Blob.Size, maxUploadSize),
 						}
 					}
 
@@ -82,8 +95,23 @@ func NewServer(id principal.Signer, storageService Service) (server.ServerView, 
 			blob.AcceptAbility,
 			server.Provide(
 				blob.Accept,
-				func(cap ucan.Capability[blob.AcceptCaveats], inv invocation.Invocation, ctx server.InvocationContext) (blob.AcceptOk, receipt.Effects, error) {
+				func(cap ucan.Capability[blob.AcceptCaveats], inv invocation.Invocation, iCtx server.InvocationContext) (blob.AcceptOk, receipt.Effects, error) {
 					log.Infof("%s z%s => %s", blob.AcceptAbility, cap.Nb().Blob.Digest.B58String(), cap.Nb().Space)
+					ctx := context.TODO()
+
+					// TODO: restrict invocation to authority (storacha service)
+
+					_, err := storageService.Blobs().Get(ctx, cap.Nb().Blob.Digest)
+					if err != nil {
+						if errors.Is(err, store.ErrNotFound) {
+							return blob.AcceptOk{}, nil, capability.Failure{
+								Name:    "AllocatedMemoryHadNotBeenWrittenTo",
+								Message: "Blob not found",
+							}
+						}
+						return blob.AcceptOk{}, nil, failure.FromError(err)
+					}
+
 					return blob.AcceptOk{}, nil, nil
 				},
 			),
